+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Enhancements
 
+- `Dune.Allowlist.Default` allows the new `Duration` module
 - Add an `:inspect_sort_maps` option for deterministic outputs
 
 ## v0.3.6 (2023-12-23)